Time: 0 ms (100.00%) | Memory: 9.8 MB (94.65%) - LeetSync

+++ 88-merge-sorted-array/merge-sorted-array.c
@@ -1,7 +1,7 @@
 void merge(int* nums1, int nums1Size, int m, int* nums2, int nums2Size, int n) {
     n--;
     m--;
-    int curi = nums1Size - 1;
+    short curi = nums1Size - 1;
     while (n >= 0 && m >= 0) {
         if (nums1[m] > nums2[n]) {
             nums1[curi] = nums1[m];
